CBOR.Int & CBOR.BigInt => CBOR.Int

+++ test/maps.js
@@ -21,9 +21,9 @@ map = map.merge(
     CBOR.Map().set(CBOR.Int(1), CBOR.String("hi")).set(CBOR.Int(5), CBOR.String("yeah")));
 assertTrue("size-3", map.length == 3);
 assertTrue("merge-0", map.get(CBOR.Int(1)).getString() == "hi");
-assertTrue("upd-0", map.update(CBOR.Int(1), CBOR.BigInt(-8n), true).getString() == "hi");
+assertTrue("upd-0", map.update(CBOR.Int(1), CBOR.Int(-8n), true).getString() == "hi");
 assertTrue("upd-1", map.get(CBOR.Int(1)).getBigInt() == -8n);
-assertTrue("upd-2", map.update(CBOR.Int(10), CBOR.BigInt(-8n), false) == null);
+assertTrue("upd-2", map.update(CBOR.Int(10), CBOR.Int(-8n), false) == null);
 assertTrue("upd-3", map.get(CBOR.Int(10)).getBigInt() == -8n);
 
 function badKey(js) {
